pyproject.toml
[project]
name = "ai-tutor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "hiredis>=3.2.1",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "openai>=1.107.0",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.10.1",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "redis>=6.4.0",
    "sqlalchemy>=2.0.43",
    "structlog>=25.4.0",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "flake8>=7.3.0",
    "mypy>=1.17.1",
    "pre-commit>=4.3.0",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.1.0",
    "pytest-cov>=7.0.0",
]
